requirements.txt
discord.py
python-dotenv
request
pillow